Tag stuff

+++ src/main/java/dev/emi/emi/api/widget/ButtonWidget.java
@@ -49,7 +49,7 @@ public class ButtonWidget extends Widget {
 	}
 
 	@Override
-	public boolean mouseClicked(double mouseX, double mouseY, int button) {
+	public boolean mouseClicked(int mouseX, int mouseY, int button) {
 		action.click(mouseX, mouseY, button);
 		MinecraftClient.getInstance().getSoundManager().play(PositionedSoundInstance.master(SoundEvents.UI_BUTTON_CLICK, 1.0f));
 		return true;
